Check for nil hostlist pointer when calling FindHosts
The pointer will return nil if hosts file is empty, as in utils.go line 65-67 when hostlist length
is zero, it will return nil instead of the empty list. In receiver function exchangeFromHosts for
outbound struct, it needs to check to see if it's nil to return false directly. Since you cannot
check for the non-exported field 'hostlist' in Hosts struct and try not to break the check,
a nil check is added instead of refactoring.